requirements.txt
flask
flask-mongoengine
flask-mongorest
flask-restplus
flask-script
requests
redis
onesignal-python

pytest
pytest-html
pytest-cov
mock
